+++ CHANGELOG.md
@@ -8,16 +8,23 @@ Todas as mudanças notáveis serão documentadas neste arquivo.
 
 ### Features
 
-## 1.0.1 (2022-01-27)
+## 1.0.1 (27-01-2022)
 
 - Adicionado conexão com banco de dados
 - Tratamento de erros
 - Middlewares "Helmet, xss, sanitize, compression, cors, httpstatus,morgan, config e http tratamento de erros.
 - Adicionado rota principal api/v1
 
-## 1.0.1 (28/01/2022)
+## 1.0.2 (29-01-2022)
 
 - Adicionado configs
 - Tratamento de Erros
 - License MIT
 - Validações
+
+## 1.0.3 (30-01-2022)
+
+- Adicionado arquivos de autenticação
+- Dependencias de autenticaçoes
+- Descrição dos middlewares
+- Limitando quantidade de acessos com senha ou email incorretos